random.c: Don't rely on libc rand() at all.  Use a Tausworthe PRNG.
Seed from urandom if possible, but if it fails, fall back on
clock_gettime(CLOCK_REALTIME).

The clock_gettime fallback uses the full range of both sec and nsec
with a syscall-per-byte.  This arrangement is somewhat better than just
using the traditional single time() call for libc rand() seeding, but it
will still be very predictable on a known platform, particularly when
the machine is idle.  At least the nsec delays between syscalls will
provide some random bits, however.

If clock_gettime fails, then we abort the caller.  Even with a PRNG,
random should not give constant values ever, and a system with no
urandom and no system clock is alien enough that I'm not going to
support it by default.