exec: Treat quoted arguments as a single argv element.
This change makes the use of nk_execute more intuitive.  Effectively
the only change is that spaces within a quoted region will not
cause a split to a new argv element.  Nested quotes are handled
properly, and any zero-length arguments produced are ignored.

The new code is also a bit easier to read.