random: Correct the getrandom syscall wrapper.
The asm-generic include should not have been used; it pulls in the
wrong system call numbers and forces the 'getrandom' syscall
to return no entropy but report success on x64 and fail with
a non-sensical errno on x86.

This is obviously bad as it means that the random output is
completely predictable if the input buffer itself does not
contain random bytes.  The only way to have detected this
problem would be via exhaustive testing.

Fix the problem, and while we're at it, check the return value
and guard against all types of short reads.  These should never
happen as we are using the urandom pool with a very short buffer
and thus should never block in the current kernel implementation,
but after the previous bug, I want to take no chances and set
no bad examples.

These new checks would have mitigated the problem by falling
back on the urandom character device.

Found by installing ndhc on an old x86 machine and noticing
strange errors caused by the old suicide() call in the getrandom
code that existed prior to the previous commit.

This issue also would never have happened if glibc actually
provided a wrapper for getrandom as it does for most system
calls.  There still isn't one, and getrandom was introduced
in 3.17, well over a year ago...